requirements.txt
requests
python-dotenv
urllib3








